+++ app/src/main/java/com/ofamosoron/dividimos/ui/composables/home/GuestCard.kt
@@ -2,13 +2,13 @@ package com.ofamosoron.dividimos.ui.composables.home
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.BoxScope
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Card
@@ -17,6 +17,8 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import com.ofamosoron.dividimos.domain.models.Guest
 import com.ofamosoron.dividimos.ui.dragAndDrop.DragTarget
@@ -24,13 +26,12 @@ import com.ofamosoron.dividimos.ui.dragAndDrop.DragTarget
 @Composable
 fun BoxScope.GuestsContainer(
     guests: List<Guest>,
-    cardClick: (String) -> Unit
+    cardClick: (String) -> Unit,
+    modifier: Modifier = Modifier
 ) {
-    LazyRow(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(14.dp)
-            .align(Alignment.BottomStart)
+    LazyColumn(
+        horizontalAlignment = Alignment.CenterHorizontally,
+        modifier = modifier.padding(end = 16.dp)
     ) {
         items(items = guests) { guest ->
             GuestCard(guest = guest) { guestUuid ->
@@ -44,17 +45,17 @@ fun BoxScope.GuestsContainer(
 @Composable
 fun GuestCard(guest: Guest, onClick: (String) -> Unit) {
     DragTarget(dataToDrop = guest, modifier = Modifier) {
-        Card(shape = RoundedCornerShape(5.dp),
-            modifier = Modifier.clickable { onClick(guest.uuid) }) {
-            Column(
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally,
-                modifier = Modifier
-                    .background(color = MaterialTheme.colorScheme.secondaryContainer)
-                    .padding(16.dp)
-            ) {
-                Text(text = guest.name, color = MaterialTheme.colorScheme.onSecondaryContainer)
-            }
+        Column(
+            verticalArrangement = Arrangement.Center,
+            horizontalAlignment = Alignment.CenterHorizontally,
+            modifier = Modifier
+                .clip(RoundedCornerShape(5.dp))
+                .background(color = MaterialTheme.colorScheme.secondaryContainer)
+                .fillMaxSize()
+                .padding(6.dp)
+                .clickable { onClick(guest.uuid) })
+        {
+            Text(text = guest.name, color = MaterialTheme.colorScheme.onSecondaryContainer)
         }
     }
 }
